trading-graph-server/pyproject.toml
[project]
name = "agent"
version = "0.1.21"
description = ""
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.66,<1.0.0",
    "langchain-openai>=0.2.14,<1.0.0", 
    "langchain-anthropic>=0.1.0,<0.4.0",
    "langchain-google-genai>=2.0.8,<3.0.0",
    "langgraph>=0.3.27,<1.0.0",
    "python-dotenv==1.0.1",
    "httpx>=0.25.0",  # For async HTTP requests
    "aioredis>=2.0.0",  # For Redis async caching support
    "h2>=4.0.0",  # For HTTP/2 protocol support
    "beautifulsoup4",
    "tenacity",
    "yfinance",
    "pandas",
    "numpy",
    "stockstats",
    "simfin",
    "finnhub-python",
    "tqdm",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "agent",
    "agent.analysts",
    "agent.controllers", 
    "agent.dataflows",
    "agent.factories",
    "agent.graph",
    "agent.graph.nodes",
    "agent.interfaces",
    "agent.judges",
    "agent.managers",
    "agent.monitoring",
    "agent.orchestrators",
    "agent.prompts",
    "agent.researchers",
    "agent.risk_mgmt",
    "agent.trader",
    "agent.utils",
    "langgraph.templates.agent"
]

[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"
"agent.analysts" = "src/agent/analysts"
"agent.controllers" = "src/agent/controllers"
"agent.dataflows" = "src/agent/dataflows"
"agent.factories" = "src/agent/factories"
"agent.graph" = "src/agent/graph"
"agent.graph.nodes" = "src/agent/graph/nodes"
"agent.interfaces" = "src/agent/interfaces"
"agent.judges" = "src/agent/judges"
"agent.managers" = "src/agent/managers"
"agent.monitoring" = "src/agent/monitoring"
"agent.orchestrators" = "src/agent/orchestrators"
"agent.prompts" = "src/agent/prompts"
"agent.researchers" = "src/agent/researchers"
"agent.risk_mgmt" = "src/agent/risk_mgmt"
"agent.trader" = "src/agent/trader"
"agent.utils" = "src/agent/utils"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]

